pyproject.toml
[project]
name = "kaggle-notifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=1.7.4.5",
    "python-dotenv>=1.2.0",
]
